Based on feedback from the community, changed the "practcl.tcl.in" template to "config.tcl.in", and it stores a key/value list instead of a stream of Tcl commands
Rewrote chunks of practcl to better handle builds in a different directory than the 
source checkout.

The kit builder now creates its own build folders for binary packages

Package objects now compute a builddir on startup

Replaced calls the the global ::project array with calls to [define get]. (Note:
make files are still welcome to write to a ::project array. That array is simply
no longer assumed to be present and populated by most of the practcl library.)