Added an AC_SUBST to capture if this extension is a Tk extension
Added a "fossil_status" command to capture data from fossil (because 
the standard fossil command redirects stdout)

Eliminated the internal mkzip implementation, and replaced with a mechanism
to download a local copy of tcllib

Added a Tcl->C encoder

Bootstrap Tcl scripts are now embedded in C rather than distributed as
a seperate file

Tweaks to the kitbuilding process

Added a "Collate_Source" method, which allows a project to generate
its dynamic sources during the compilation sequence, instead of 
during initialization of the object

Created a class to handle behaviors particular to autoconf based
projects, and split that behavior from the library class.

Created a new "project" class to act as a high level abstraction for
projects which may implement both a kit and a dynamic library

Eliminated the need to have a copy of odie in order to generate a basic kit.

Added behaviors to exercise a practcl style make.tcl script if present within
a project.