Adding support for cross compiling. Practcl now looks for a "CONFIG_SITE" element in the master project's define array. If present, practcl adapts itself to perform a cross compile.
If ::practcl::config.tcl is run, this entry will be automatically created
if a config.site file is in the directory of the master project build.

CONFIG_SITE points to a file containing a sh compadible description of the
cross compiler environment. Practcl will also read and respond to flags in
this file